get rid of NonInteractive base class
