+++ lib/teachers_pet/actions/open_issue.rb
@@ -1,6 +1,6 @@
 module TeachersPet
   module Actions
-    class OpenIssue < NonInteractive
+    class OpenIssue < Base
       def read_info
         @repository = self.options[:repository]
         @organization = self.options[:organization]
